Remove : api 경로 이동
